Split AppModule into lazy-loaded Modules - DesktopModule - ControlCenterModule - AccountModule
