+++ src/app/app.component.spec.ts
@@ -3,8 +3,8 @@ import {AppComponent} from './app.component';
 import {DesktopComponent} from './desktop/desktop.component';
 import {DesktopMenuComponent} from './desktop/desktop-menu/desktop-menu.component';
 import {DesktopStartmenuComponent} from './desktop/desktop-startmenu/desktop-startmenu.component';
-import {PXtoViewWidthPipe} from './pxto-view-width.pipe';
-import {PXtoViewHeightPipe} from './pxto-view-height.pipe';
+import {PXtoViewWidthPipe} from '../core/px-to/pxto-view-width.pipe';
+import {PXtoViewHeightPipe} from '../core/px-to/pxto-view-height.pipe';
 import {WindowManagerComponent} from './desktop/window-manager/window-manager.component';
 import {WindowFrameComponent} from './desktop/window/window-frame.component';
 import {WebsocketService} from './websocket.service';
