+++ src/app/control-center/control-center.module.ts
@@ -1,6 +1,5 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {ControlCenterComponent} from './control-center.component';
 import {ControlCenterSidebarComponent} from './control-center-sidebar/control-center-sidebar.component';
 import {ControlCenterSidebarMenuComponent} from './control-center-sidebar-menu/control-center-sidebar-menu.component';
@@ -15,54 +14,11 @@ import {
   ControlCenterChangelogPageComponent
 } from './control-center-changelog-page/control-center-changelog-page.component';
 import {ControlCenterSoundPageComponent} from './control-center-sound-page/control-center-sound-page.component';
-import {RouterModule, Routes} from '@angular/router';
-import {ControlCenterGuard} from './control-center.guard';
-import {ControlCenterService} from './control-center.service';
-import {
-  ControlCenterDevicePageHardwareResolver
-} from './control-center-device-page/control-center-device-page-hardware.resolver';
 import {
   ControlCenterInventoryPageComponent
 } from './control-center-inventory-page/control-center-inventory-page.component';
-import {
-  ControlCenterInventoryPageItemsResolver
-} from './control-center-inventory-page/control-center-inventory-page-items.resolver';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-
-
-const routes: Routes = [
-  {
-    path: '',
-    component: ControlCenterComponent,
-    canActivate: [ControlCenterGuard],
-    resolve: {
-      devices: ControlCenterService
-    },
-    data: {animation: 'control-center'},
-    children: [
-      {
-        path: 'device',
-        component: ControlCenterDevicePageComponent,
-        runGuardsAndResolvers: 'paramsOrQueryParamsChange',
-        resolve: {hardware: ControlCenterDevicePageHardwareResolver}
-      },
-      {
-        path: 'create-device',
-        component: ControlCenterCreateDevicePageComponent,
-        resolve: {inventoryItems: ControlCenterInventoryPageItemsResolver}
-      },
-      {
-        path: 'inventory',
-        component: ControlCenterInventoryPageComponent,
-        resolve: {items: ControlCenterInventoryPageItemsResolver}
-      },
-      {path: 'settings', component: ControlCenterSettingsPageComponent},
-      {path: 'sound', component: ControlCenterSoundPageComponent},
-      {path: 'changelog', component: ControlCenterChangelogPageComponent}
-    ]
-  }
-];
-
+import {ControlCenterRoutingModule} from "./control-center-routing.module";
 
 @NgModule({
   declarations: [
@@ -78,8 +34,7 @@ const routes: Routes = [
   ],
   imports: [
     CommonModule,
-    BrowserAnimationsModule,
-    RouterModule.forChild(routes),
+    ControlCenterRoutingModule,
     ReactiveFormsModule,
     FormsModule
   ]
