--- src/app/design/button/button.component.html
@@ -1,1 +1,0 @@
-<ng-content select="input[type=submit]"></ng-content>
