--- src/app/design/button/button.component.ts
@@ -1,9 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-  selector: 'app-button',
-  templateUrl: './button.component.html',
-  styleUrls: ['./button.component.scss']
-})
-export class ButtonComponent {
-}
