--- src/app/design/form-group/form-group.component.html
@@ -1,1 +1,0 @@
-<ng-content></ng-content>
