--- src/app/design/form-group/form-group.component.ts
@@ -1,9 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-  selector: 'app-form-group',
-  templateUrl: './form-group.component.html',
-  styleUrls: ['./form-group.component.scss']
-})
-export class FormGroupComponent {
-}
