--- src/app/design/text-field/text-field.component.html
@@ -1,4 +1,0 @@
-<ng-content select="input"></ng-content>
-<ng-content select="textarea"></ng-content>
-
-<!-- TODO: Autocomplete -->
