--- src/app/design/text-field/text-field.component.spec.ts
@@ -1,25 +1,0 @@
-import {ComponentFixture, TestBed, waitForAsync} from '@angular/core/testing';
-
-import {TextFieldComponent} from './text-field.component';
-
-describe('TextFieldComponent', () => {
-  let component: TextFieldComponent;
-  let fixture: ComponentFixture<TextFieldComponent>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [TextFieldComponent]
-    })
-      .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(TextFieldComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
