--- src/app/design/text-field/text-field.component.ts
@@ -1,9 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-  selector: 'app-text-field',
-  templateUrl: './text-field.component.html',
-  styleUrls: ['./text-field.component.scss']
-})
-export class TextFieldComponent {
-}
