+++ src/app/desktop/desktop-startmenu/desktop-startmenu.component.spec.ts
@@ -5,8 +5,8 @@ import {FormsModule} from '@angular/forms';
 import {HttpClientModule} from '@angular/common/http';
 import {RouterTestingModule} from '@angular/router/testing';
 import {DesktopComponent} from '../desktop.component';
-import {PXtoViewHeightPipe} from '../../pxto-view-height.pipe';
-import {PXtoViewWidthPipe} from '../../pxto-view-width.pipe';
+import {PXtoViewHeightPipe} from '../../../core/px-to/pxto-view-height.pipe';
+import {PXtoViewWidthPipe} from '../../../core/px-to/pxto-view-width.pipe';
 import {WindowManagerComponent} from '../window-manager/window-manager.component';
 import {WindowFrameComponent} from '../window/window-frame.component';
 import {DesktopMenuComponent} from '../desktop-menu/desktop-menu.component';
