+++ src/app/desktop/desktop.component.spec.ts
@@ -4,8 +4,8 @@ import {DesktopComponent} from './desktop.component';
 import {DesktopMenuComponent} from './desktop-menu/desktop-menu.component';
 import {HttpClientModule} from '@angular/common/http';
 import {RouterTestingModule} from '@angular/router/testing';
-import {PXtoViewHeightPipe} from '../pxto-view-height.pipe';
-import {PXtoViewWidthPipe} from '../pxto-view-width.pipe';
+import {PXtoViewHeightPipe} from '../../core/px-to/pxto-view-height.pipe';
+import {PXtoViewWidthPipe} from '../../core/px-to/pxto-view-width.pipe';
 import {DesktopStartmenuComponent} from './desktop-startmenu/desktop-startmenu.component';
 import {FormsModule} from '@angular/forms';
 import {WindowManagerComponent} from './window-manager/window-manager.component';
