+++ src/app/desktop/windows/file-manager/file-manager.component.spec.ts
@@ -5,7 +5,7 @@ import {FormsModule} from '@angular/forms';
 import {WebsocketService} from '../../../websocket.service';
 import {emptyWindowDelegate, webSocketMock, windowManagerMock} from '../../../test-utils';
 import {WindowManager} from '../../window-manager/window-manager';
-import {ContextMenuModule} from "../../../design/context-menu/context-menu.module";
+import {ContextMenuModule} from "../../../../core/components/context-menu/context-menu.module";
 
 describe('FileManagerComponent', () => {
   let component: FileManagerComponent;
