+++ src/core/components/context-menu/context-menu-content/context-menu-content.component.scss
