+++ src/core/components/context-menu/context-menu/context-menu.component.scss
